// consulting-gpt-frontend/types/comparison-axis.ts
export interface ComparisonAxis {
  axisName: string;
  extreme1: string;
  extreme2: string;
  weight: number;
  id: number;
  technologyId: number;
}

export interface ComparisonAxisRaw {
  axis_name: string;
  extreme1: string;
  extreme2: string;
  weight: number;
  id: number;
  technology_id: number;
}